+++ romancal/ramp_fitting/ramp_fit_step.py
@@ -180,11 +180,11 @@ class RampFitStep(RomanStep):
             log.info('All pixels are saturated. Returning a zeroed-out image.')
 
             # Image info order is: data, dq, var_poisson, var_rnoise, err
-            image_info = (np.zeros(input_model.data.shape[2:]),
+            image_info = (np.zeros(input_model.data.shape[2:], dtype=input_model.data.dtype),
                           input_model.pixeldq | input_model.groupdq[0][0] | dqflags.group['SATURATED'],
-                          np.zeros(input_model.err.shape[2:]),
-                          np.zeros(input_model.err.shape[2:]),
-                          np.zeros(input_model.err.shape[2:]))
+                          np.zeros(input_model.err.shape[2:], dtype=input_model.err.dtype),
+                          np.zeros(input_model.err.shape[2:], dtype=input_model.err.dtype),
+                          np.zeros(input_model.err.shape[2:], dtype=input_model.err.dtype))
 
         out_model = create_image_model(input_model, image_info)
         out_model.meta.cal_step.ramp_fit = 'COMPLETE'
